Revision 9.6.1. Refactoring stuff. * Extracted an enum for PlayerInteraction type. * Formatted many classes. * Added 'javastory.config' package. * Floating-point server rates! It's funny, people /never/ thought of making a <1x server. Would be so awesome! * Also, rates stored in the database. Stupid config file is stupid. * Movement parsing uses PacketReader.readVector() instead of reading point components separately. Shouldn't be faster, however so much prettier! * Added 'javastory.game.data' package for data provider classes. Should move all of them later.

+++ Core/src/main/java/javastory/channel/life/Monster.java
@@ -319,7 +319,7 @@ public class Monster extends AbstractLoadedLife {
 	}
 
 	public final ChannelCharacter killBy(final ChannelCharacter killer) {
-		int totalBaseExp = (Math.min(Integer.MAX_VALUE, (getMobExp() * ChannelServer.getInstance().getExpRate())));
+		int totalBaseExp = (Math.min(Integer.MAX_VALUE, (int) (getMobExp() * ChannelServer.getInstance().getExpRate())));
 		AttackerEntry highest = null;
 		int highdamage = 0;
 		for (final AttackerEntry attackEntry : attackers) {
